Conflicts:
	browser/components/sessionstore/src/SessionStore.jsm
	browser/devtools/commandline/BuiltinCommands.jsm
	browser/devtools/scratchpad/scratchpad.js
	configure.in
	content/base/src/nsXSSUtils.cpp
	content/media/AudioStream.cpp
	content/media/MediaDecoder.cpp
	content/media/gstreamer/GStreamerFormatHelper.cpp
	content/media/gstreamer/GStreamerFormatHelper.h
	content/media/gstreamer/GStreamerFunctionList.h
	content/media/gstreamer/GStreamerReader.cpp
	content/media/gstreamer/GStreamerReader.h
	modules/libpref/src/init/all.js
	netwerk/base/public/security-prefs.js
	security/manager/boot/src/nsSTSPreloadList.inc

+++ media/webrtc/trunk/webrtc/modules/video_coding/main/source/content_metrics_processing.cc
@@ -8,12 +8,12 @@
  *  be found in the AUTHORS file in the root of the source tree.
  */
 
-#include "modules/video_coding/main/source/content_metrics_processing.h"
+#include "webrtc/modules/video_coding/main/source/content_metrics_processing.h"
 
 #include <math.h>
 
-#include "modules/interface/module_common_types.h"
-#include "modules/video_coding/main/interface/video_coding_defines.h"
+#include "webrtc/modules/interface/module_common_types.h"
+#include "webrtc/modules/video_coding/main/interface/video_coding_defines.h"
 
 namespace webrtc {
 //////////////////////////////////
@@ -21,7 +21,7 @@ namespace webrtc {
 //////////////////////////////////
 
 VCMContentMetricsProcessing::VCMContentMetricsProcessing()
-    : recursive_avg_factor_(1 / 150.0f),  // matched to  30fps.
+    : recursive_avg_factor_(1 / 150.0f),  // matched to  15fps.
       frame_cnt_uniform_avg_(0),
       avg_motion_level_(0.0f),
       avg_spatial_level_(0.0f) {
@@ -43,7 +43,7 @@ int VCMContentMetricsProcessing::Reset() {
   return VCM_OK;
 }
 
-void VCMContentMetricsProcessing::UpdateFrameRate(uint32_t frameRate) {
+void VCMContentMetricsProcessing::UpdateFrameRate(float frameRate) {
   // Update factor for recursive averaging.
   recursive_avg_factor_ = static_cast<float> (1000.0f) /
       static_cast<float>(frameRate *  kQmMinIntervalMs);
@@ -122,4 +122,4 @@ void VCMContentMetricsProcessing::UpdateRecursiveAvg(
       recursive_avg_->motion_magnitude +
       recursive_avg_factor_ * contentMetrics->motion_magnitude;
 }
-}  // end of namespace
+}  // namespace
